move fetcher into serializers module
