--- app/services/pipeline_service/serializer_fetcher.rb
@@ -1,16 +1,0 @@
-module PipelineService
-  module SerializerFetcher
-    def self.fetch(object: )
-      case object
-      when Enrollment, DesignerEnrollment, ObserverEnrollment, StudentEnrollment, TeacherEnrollment
-        PipelineService::Serializers::Enrollment
-      when Submission
-        PipelineService::Serializers::Submission
-      when User
-        PipelineService::Serializers::User
-      else
-        raise NameError.new("Could not find the serializer for #{object}")
-      end
-    end
-  end
-end
