+++ spec/services/pipeline_service/pipeline_client_spec.rb
@@ -1,10 +1,11 @@
 describe PipelineService::PipelineClient do
   let(:endpoint_instance) { double('endpoint_instance', call: nil)}
   let(:endpoint_class) { double('endpoint_class', new: endpoint_instance) }
+  let(:serializer) {double('serializer class', new: double('serializer instance', call: nil))}
 
   subject do
     described_class.new(
-      object: nil,
+      object: Enrollment.new,
       noun: '',
       id: 1,
       endpoint: endpoint_class
@@ -13,6 +14,13 @@ describe PipelineService::PipelineClient do
 
   it 'posts to the endpoint' do
     expect(endpoint_instance).to receive(:call).and_return(nil)
+    subject.call
+  end
+
+  it 'looks up the serializer' do
+    expect(PipelineService::Serializers::Fetcher)
+      .to receive(:fetch)
+      .and_return(serializer)
     subject.call
   end
